Refactor CLI and Streamlit modules
