+++ src/autoresearch/main.py
@@ -4,18 +4,14 @@ from __future__ import annotations
 
 import sys
 import os
-import difflib
-from typing import Optional, List, Sequence, Any
-import click
+from typing import Optional, Any
 
 import typer
 from rich.console import Console
-from rich.table import Table
 from rich.prompt import Prompt
 from rich.progress import Progress
 from .mcp_interface import create_server
 from .monitor import monitor_app
-from datetime import datetime
 import time
 
 from .config import ConfigLoader, ConfigModel
@@ -24,8 +20,7 @@ from .orchestration.state import QueryState
 from .output_format import OutputFormatter
 from .logging_utils import configure_logging
 from .storage import StorageManager
-from .storage_backup import BackupManager, BackupConfig
-from .errors import BackupError, StorageError
+from .errors import StorageError
 from .cli_utils import (
     console,
     print_success,
@@ -44,59 +39,8 @@ from .cli_utils import (
     sparql_query_cli as _cli_sparql,
 )
 from .error_utils import get_error_info, format_error_for_cli
-
-
-def find_similar_commands(
-    command: str, valid_commands: Sequence[str], threshold: float = 0.6
-) -> List[str]:
-    """Find similar commands to the given command.
-
-    Args:
-        command: The command to find similar commands for
-        valid_commands: The list of valid commands
-        threshold: The similarity threshold (0.0 to 1.0)
-
-    Returns:
-        A list of similar commands
-    """
-    # Use difflib to find similar commands
-    matches = difflib.get_close_matches(
-        command,
-        valid_commands,
-        n=3,
-        cutoff=threshold,
-    )
-    return list(matches)
-
-
-def handle_command_not_found(ctx: typer.Context, command: str) -> None:
-    """Handle command not found errors with helpful suggestions.
-
-    Args:
-        ctx: The Typer context
-        command: The command that was not found
-    """
-    print_error(f"Command '{command}' not found.")
-
-    # Get all available commands
-    available_commands: List[str] = []
-    if isinstance(ctx.command, click.Group):
-        for command_obj in ctx.command.commands.values():
-            if command_obj.name:
-                available_commands.append(command_obj.name)
-
-    # Find similar commands
-    similar_commands = find_similar_commands(command, available_commands)
-
-    if similar_commands:
-        print_info("Did you mean:", symbol=False)
-        for cmd in similar_commands:
-            print_command_example(cmd)
-
-    console.print(
-        "\nRun [cyan]autoresearch --help[/cyan] to see all available commands."
-    )
-    raise typer.Exit(code=1)
+from .cli_helpers import handle_command_not_found
+from .cli_backup import backup_app
 
 
 app = typer.Typer(
@@ -489,6 +433,7 @@ def search(
 
 config_app = typer.Typer(help="Configuration management commands")
 app.add_typer(config_app, name="config")
+app.add_typer(backup_app, name="backup")
 
 
 @config_app.callback(invoke_without_command=True)
@@ -775,8 +720,6 @@ def serve_a2a(
 
         # Keep the main thread running until interrupted
         try:
-            import time
-
             while True:
                 time.sleep(1)
         except KeyboardInterrupt:
@@ -784,432 +727,6 @@ def serve_a2a(
             console.print("[bold yellow]Server stopped[/bold yellow]")
     except Exception as e:
         console.print(f"[bold red]Error starting A2A server:[/bold red] {str(e)}")
-
-
-backup_app = typer.Typer(help="Backup and restore operations")
-app.add_typer(backup_app, name="backup")
-
-
-@backup_app.command("create")
-def backup_create(
-    backup_dir: Optional[str] = typer.Option(
-        None,
-        "--dir",
-        "-d",
-        help="Directory where backups will be stored. Defaults to 'backups' or value from config.",
-    ),
-    compress: bool = typer.Option(
-        True,
-        "--compress/--no-compress",
-        help="Whether to compress the backup. Default is to compress.",
-    ),
-    max_backups: int = typer.Option(
-        5,
-        "--max-backups",
-        "-m",
-        help="Maximum number of backups to keep. Default is 5.",
-    ),
-    retention_days: int = typer.Option(
-        30,
-        "--retention-days",
-        "-r",
-        help="Maximum age of backups in days. Default is 30.",
-    ),
-) -> None:
-    """Create a backup of the storage system.
-
-    This command creates a backup of the DuckDB database and RDF store.
-    The backup can be compressed to save space, and a rotation policy
-    can be applied to limit the number and age of backups.
-
-    Examples:
-        # Create a backup with default settings
-        autoresearch backup create
-
-        # Create a backup in a specific directory
-        autoresearch backup create --dir ~/autoresearch-backups
-
-        # Create an uncompressed backup
-        autoresearch backup create --no-compress
-
-        # Create a backup with custom retention settings
-        autoresearch backup create --max-backups 10 --retention-days 60
-    """
-    console = Console()
-
-    try:
-        # Create backup configuration
-        config = BackupConfig(
-            backup_dir=backup_dir or "backups",
-            compress=compress,
-            max_backups=max_backups,
-            retention_days=retention_days,
-        )
-
-        # Show progress
-        with Progress() as progress:
-            task = progress.add_task("[green]Creating backup...", total=1)
-
-            # Create backup
-            backup_info = BackupManager.create_backup(
-                backup_dir=backup_dir, compress=compress, config=config
-            )
-
-            progress.update(task, completed=1)
-
-        # Show backup info
-        console.print("[bold green]Backup created successfully:[/bold green]")
-        console.print(f"  Path: {backup_info.path}")
-        console.print(f"  Timestamp: {backup_info.timestamp}")
-        console.print(f"  Size: {_format_size(backup_info.size)}")
-        console.print(f"  Compressed: {'Yes' if backup_info.compressed else 'No'}")
-
-    except BackupError as e:
-        console.print(f"[bold red]Error creating backup:[/bold red] {str(e)}")
-        if hasattr(e, "context") and "suggestion" in e.context:
-            console.print(f"[yellow]Suggestion:[/yellow] {e.context['suggestion']}")
-        raise typer.Exit(code=1)
-
-
-@backup_app.command("restore")
-def backup_restore(
-    backup_path: str = typer.Argument(
-        ...,
-        help="Path to the backup to restore. Use 'list' command to see available backups.",
-    ),
-    target_dir: Optional[str] = typer.Option(
-        None,
-        "--dir",
-        "-d",
-        help="Directory where the backup will be restored. Defaults to a timestamped directory.",
-    ),
-    force: bool = typer.Option(
-        False,
-        "--force",
-        "-f",
-        help="Force restore without confirmation.",
-    ),
-) -> None:
-    """Restore a backup of the storage system.
-
-    This command restores a backup of the DuckDB database and RDF store.
-    The backup can be restored to a specific directory, or to a timestamped
-    directory by default.
-
-    Examples:
-        # Restore a specific backup (first list backups to find the path)
-        autoresearch backup restore backups/autoresearch-backup-2023-01-01-120000.tar.gz
-
-        # Restore to a specific directory
-        autoresearch backup restore backups/autoresearch-backup-2023-01-01-120000.tar.gz --dir ~/restored-data
-
-        # Force restore without confirmation
-        autoresearch backup restore backups/autoresearch-backup-2023-01-01-120000.tar.gz --force
-    """
-    console = Console()
-
-    try:
-        # Confirm restore if not forced
-        if not force:
-            console.print(
-                "[bold yellow]Warning:[/bold yellow] Restoring a backup will create new database files."
-            )
-            console.print(
-                "The original files will not be modified, but you will need to configure"
-            )
-            console.print(
-                "the application to use the restored files if you want to use them."
-            )
-            confirm = Prompt.ask(
-                "Are you sure you want to restore this backup?",
-                choices=["y", "n"],
-                default="n",
-            )
-            if confirm.lower() != "y":
-                console.print("Restore cancelled.")
-                return
-
-        # Show progress
-        with Progress() as progress:
-            task = progress.add_task("[green]Restoring backup...", total=1)
-
-            # Restore backup
-            restored_paths = BackupManager.restore_backup(
-                backup_path=backup_path, target_dir=target_dir
-            )
-
-            progress.update(task, completed=1)
-
-        # Show restored paths
-        console.print("[bold green]Backup restored successfully:[/bold green]")
-        console.print(f"  DuckDB database: {restored_paths['db_path']}")
-        console.print(f"  RDF store: {restored_paths['rdf_path']}")
-        console.print(
-            "\nTo use the restored files, update your configuration to point to these paths."
-        )
-
-    except BackupError as e:
-        console.print(f"[bold red]Error restoring backup:[/bold red] {str(e)}")
-        if hasattr(e, "context") and "suggestion" in e.context:
-            console.print(f"[yellow]Suggestion:[/yellow] {e.context['suggestion']}")
-        raise typer.Exit(code=1)
-
-
-@backup_app.command("list")
-def backup_list(
-    backup_dir: Optional[str] = typer.Option(
-        None,
-        "--dir",
-        "-d",
-        help="Directory containing backups. Defaults to 'backups' or value from config.",
-    ),
-    limit: int = typer.Option(
-        10,
-        "--limit",
-        "-n",
-        help="Maximum number of backups to list. Default is 10.",
-    ),
-) -> None:
-    """List available backups.
-
-    This command lists all backups in the specified directory,
-    showing their timestamp, size, and compression status.
-
-    Examples:
-        # List backups with default settings
-        autoresearch backup list
-
-        # List backups in a specific directory
-        autoresearch backup list --dir ~/autoresearch-backups
-
-        # List more backups than the default limit
-        autoresearch backup list --limit 20
-    """
-    console = Console()
-
-    try:
-        # List backups
-        backups = BackupManager.list_backups(backup_dir)
-
-        if not backups:
-            console.print(f"No backups found in {backup_dir or 'backups'}")
-            return
-
-        # Create table
-        table = Table(title=f"Backups in {backup_dir or 'backups'}")
-        table.add_column("Timestamp", style="cyan")
-        table.add_column("Path", style="green")
-        table.add_column("Size", style="magenta")
-        table.add_column("Compressed", style="yellow")
-
-        # Add rows
-        for backup in backups[:limit]:
-            table.add_row(
-                backup.timestamp.strftime("%Y-%m-%d %H:%M:%S"),
-                backup.path,
-                _format_size(backup.size),
-                "Yes" if backup.compressed else "No",
-            )
-
-        # Show table
-        console.print(table)
-
-        if len(backups) > limit:
-            console.print(
-                f"Showing {limit} of {len(backups)} backups. Use --limit to show more."
-            )
-
-    except BackupError as e:
-        console.print(f"[bold red]Error listing backups:[/bold red] {str(e)}")
-        if hasattr(e, "context") and "suggestion" in e.context:
-            console.print(f"[yellow]Suggestion:[/yellow] {e.context['suggestion']}")
-        raise typer.Exit(code=1)
-
-
-@backup_app.command("schedule")
-def backup_schedule(
-    interval_hours: int = typer.Option(
-        24,
-        "--interval",
-        "-i",
-        help="Interval between backups in hours. Default is 24 (daily).",
-    ),
-    backup_dir: Optional[str] = typer.Option(
-        None,
-        "--dir",
-        "-d",
-        help="Directory where backups will be stored. Defaults to 'backups' or value from config.",
-    ),
-    compress: bool = typer.Option(
-        True,
-        "--compress/--no-compress",
-        help="Whether to compress the backup. Default is to compress.",
-    ),
-    max_backups: int = typer.Option(
-        5,
-        "--max-backups",
-        "-m",
-        help="Maximum number of backups to keep. Default is 5.",
-    ),
-    retention_days: int = typer.Option(
-        30,
-        "--retention-days",
-        "-r",
-        help="Maximum age of backups in days. Default is 30.",
-    ),
-) -> None:
-    """Schedule periodic backups.
-
-    This command schedules periodic backups of the storage system.
-    The backups will be created at the specified interval, and a
-    rotation policy will be applied to limit the number and age of backups.
-
-    Note: This command starts a background process that will continue
-    running until the application is stopped.
-    """
-    console = Console()
-
-    try:
-        # Schedule backups
-        BackupManager.schedule_backup(
-            backup_dir=backup_dir,
-            interval_hours=interval_hours,
-            compress=compress,
-            max_backups=max_backups,
-            retention_days=retention_days,
-        )
-
-        console.print(
-            f"[bold green]Scheduled backups every {interval_hours} hours to {backup_dir or 'backups'}[/bold green]"
-        )
-        console.print("The first backup will be created immediately.")
-        console.print(
-            "Press Ctrl+C to stop the application and cancel scheduled backups."
-        )
-
-        # Keep the application running
-        try:
-            while True:
-                time.sleep(1)
-        except KeyboardInterrupt:
-            console.print("[bold yellow]Stopping scheduled backups...[/bold yellow]")
-            BackupManager.stop_scheduled_backups()
-            console.print("[bold green]Scheduled backups stopped.[/bold green]")
-
-    except BackupError as e:
-        console.print(f"[bold red]Error scheduling backups:[/bold red] {str(e)}")
-        if hasattr(e, "context") and "suggestion" in e.context:
-            console.print(f"[yellow]Suggestion:[/yellow] {e.context['suggestion']}")
-        raise typer.Exit(code=1)
-
-
-@backup_app.command("recover")
-def backup_recover(
-    timestamp: str = typer.Argument(
-        ...,
-        help="Target timestamp for point-in-time recovery (YYYY-MM-DD HH:MM:SS).",
-    ),
-    backup_dir: Optional[str] = typer.Option(
-        None,
-        "--dir",
-        "-d",
-        help="Directory containing backups. Defaults to 'backups' or value from config.",
-    ),
-    target_dir: Optional[str] = typer.Option(
-        None,
-        "--target-dir",
-        "-t",
-        help="Directory where the backup will be restored. Defaults to a timestamped directory.",
-    ),
-    force: bool = typer.Option(
-        False,
-        "--force",
-        "-f",
-        help="Force recovery without confirmation.",
-    ),
-) -> None:
-    """Perform point-in-time recovery.
-
-    This command restores the storage system to a specific point in time
-    by finding the backup closest to the specified timestamp and restoring it.
-    """
-    console = Console()
-
-    try:
-        # Parse timestamp
-        try:
-            target_time = datetime.strptime(timestamp, "%Y-%m-%d %H:%M:%S")
-        except ValueError:
-            console.print(
-                "[bold red]Error:[/bold red] Invalid timestamp format. Use YYYY-MM-DD HH:MM:SS."
-            )
-            raise typer.Exit(code=1)
-
-        # Confirm recovery if not forced
-        if not force:
-            console.print(
-                "[bold yellow]Warning:[/bold yellow] Point-in-time recovery will create new database files."
-            )
-            console.print(
-                "The original files will not be modified, but you will need to configure"
-            )
-            console.print(
-                "the application to use the recovered files if you want to use them."
-            )
-            confirm = Prompt.ask(
-                "Are you sure you want to perform point-in-time recovery?",
-                choices=["y", "n"],
-                default="n",
-            )
-            if confirm.lower() != "y":
-                console.print("Recovery cancelled.")
-                return
-
-        # Show progress
-        with Progress() as progress:
-            task = progress.add_task(
-                "[green]Performing point-in-time recovery...", total=1
-            )
-
-            # Perform recovery
-            restored_paths = BackupManager.restore_point_in_time(
-                backup_dir=backup_dir or "backups",
-                target_time=target_time,
-                target_dir=target_dir,
-            )
-
-            progress.update(task, completed=1)
-
-        # Show restored paths
-        console.print(
-            "[bold green]Point-in-time recovery completed successfully:[/bold green]"
-        )
-        console.print(f"  Target time: {target_time}")
-        console.print(f"  DuckDB database: {restored_paths['db_path']}")
-        console.print(f"  RDF store: {restored_paths['rdf_path']}")
-        console.print(
-            "\nTo use the recovered files, update your configuration to point to these paths."
-        )
-
-    except BackupError as e:
-        console.print(
-            f"[bold red]Error performing point-in-time recovery:[/bold red] {str(e)}"
-        )
-        if hasattr(e, "context") and "suggestion" in e.context:
-            console.print(f"[yellow]Suggestion:[/yellow] {e.context['suggestion']}")
-        raise typer.Exit(code=1)
-
-
-def _format_size(size_bytes: int) -> str:
-    """Format a size in bytes as a human-readable string."""
-    if size_bytes < 1024:
-        return f"{size_bytes} B"
-    elif size_bytes < 1024 * 1024:
-        return f"{size_bytes / 1024:.1f} KB"
-    elif size_bytes < 1024 * 1024 * 1024:
-        return f"{size_bytes / (1024 * 1024):.1f} MB"
-    else:
-        return f"{size_bytes / (1024 * 1024 * 1024):.1f} GB"
 
 
 @app.command("completion")
