+++ tests/unit/test_coalition_execution.py
@@ -1,5 +1,7 @@
 from autoresearch.orchestration.orchestrator import Orchestrator
 from autoresearch.config import ConfigModel
+from autoresearch.agents.registry import AgentFactory
+
 
 class DummyAgent:
     def __init__(self, name, record):
@@ -14,18 +16,24 @@ class DummyAgent:
         return {}
 
 
-def test_coalition_agents_run_together(monkeypatch):
+def test_coalition_agents_run_together(monkeypatch, tmp_path):
     record = []
+    AgentFactory._registry.clear()
 
     def get_agent(name):
         return DummyAgent(name, record)
+
+    AgentFactory.register("A", DummyAgent)
+    AgentFactory.register("B", DummyAgent)
 
     cfg = ConfigModel(agents=["team"], loops=1, coalitions={"team": ["A", "B"]})
 
     monkeypatch.setattr(
         "autoresearch.orchestration.orchestrator.AgentFactory.get", get_agent
     )
+    monkeypatch.setenv("AUTORESEARCH_RELEASE_METRICS", str(tmp_path / "rel.json"))
+    monkeypatch.setenv("AUTORESEARCH_QUERY_TOKENS", str(tmp_path / "qt.json"))
 
     Orchestrator.run_query("q", cfg)
 
-    assert record == ["A", "B"]
+    assert set(record[:2]) == {"A", "B"}
