+++ tests/unit/test_streamlit_utils.py
@@ -17,7 +17,7 @@ sys.modules.setdefault("matplotlib", fake_matplotlib)
 sys.modules.setdefault("matplotlib.pyplot", ModuleType("pyplot"))
 sys.modules.setdefault("psutil", ModuleType("psutil"))
 
-from autoresearch.streamlit_app import (  # noqa: E402
+from autoresearch.config_utils import (  # noqa: E402
     save_config_to_toml,
     apply_preset,
 )
